Move internal LayerTextEditor to usdQtEditors, replace redundant class
